+++ lib/dashboard_grid.dart
@@ -1,3 +1,4 @@
 export 'src/dashboard.dart';
 export 'src/dashboard_grid.dart';
 export 'src/dashboard_widget.dart';
+export 'src/table/table_cell_decoration.dart';